+++ src/Storage/src/Bootloader/StorageBootloader.php
@@ -2,21 +2,22 @@
 
 declare(strict_types=1);
 
-namespace Spiral\Bootloader\Storage;
+namespace Spiral\Storage\Bootloader;
 
 use Spiral\Boot\Bootloader\Bootloader;
 use Spiral\Boot\Exception\EnvironmentException;
-use Spiral\Bootloader\Distribution\DistributionBootloader;
 use Spiral\Config\ConfiguratorInterface;
 use Spiral\Core\Container;
 use Spiral\Core\Exception\Container\NotFoundException;
+use Spiral\Distribution\Bootloader\DistributionBootloader;
+use Spiral\Distribution\DistributionInterface as CdnInterface;
+use Spiral\Storage\Bucket;
 use Spiral\Storage\BucketFactory;
 use Spiral\Storage\BucketFactoryInterface;
+use Spiral\Storage\BucketInterface;
+use Spiral\Storage\Config\StorageConfig;
 use Spiral\Storage\Storage;
 use Spiral\Storage\StorageInterface;
-use Spiral\Storage\Bucket;
-use Spiral\Storage\BucketInterface;
-use Spiral\Distribution\DistributionInterface as CdnInterface;
 
 class StorageBootloader extends Bootloader
 {
@@ -55,7 +56,7 @@ class StorageBootloader extends Bootloader
                             . 'Please make sure that bootloader %s is added in your application';
                         $message = \sprintf($message, $name, DistributionBootloader::class);
 
-                        throw new EnvironmentException($message, (int)$e->getCode(), $e);
+                        throw new EnvironmentException($message, $e->getCode(), $e);
                     }
 
                     $resolver = $cdn->resolver($distributions[$name]);